feat: replace full-scan lexical search with FTS5 MATCH + BM25
Core search rewrite: sqlite_search.py now queries the lexical_fts virtual
table with FTS5 MATCH + bm25() scoring instead of loading all index_entries
and computing IDF in Python. Visibility filtering remains in Python.

Key changes:
- Search uses FTS5 inverted index for token matching and BM25 for scoring
- Scores change from IDF integers to BM25 floats (positive, higher = better)
- matched_tokens reconstructed post-hoc by intersecting query with result text
- Container_ref filtering handled by Python visibility check (not SQL) to
  preserve cross-container visibility for public items
- Batch commit path (_persist_process_result_in_session) now also inserts
  lexical entries into the FTS5 table
- Downstream score types updated: QueryResultItem.score, RetrievalTraceHit.score,
  QueryResultResponse.score all changed from int to float
- Routing scoring adapted for BM25 float scores: removed int() truncations,
  adjusted injection thresholds, added quality floor for higher-level penalty
- 4 new FTS5 search tests: BM25 scores, container scoping, safety, ranking

+++ core/models.py
@@ -173,7 +173,7 @@ class QueryRuntimeContext:
 @dataclass(frozen=True)
 class QueryResultItem:
     result_kind: str
-    score: int
+    score: float
     evidence: list[EvidenceReference]
     result_id: str | None = None
     memory_object_id: str | None = None
@@ -228,7 +228,7 @@ class RetrievalTraceHit:
     index_entry_id: str
     index_type: str
     text_view_name: str
-    score: int
+    score: float
     matched_tokens: tuple[str, ...]
     provider_name: str | None = None
     provider_version: str | None = None
